don't warn about `window.fetch` during hydration (#5041)
* failing test for #5031

* move fetch logic into separate module - fixes #5031

* changeset

* remove .only
