+++ .changeset/young-scissors-collect.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': patch
+---
+
+don't warn about window.fetch during hydration
