+++ packages/kit/test/apps/basics/src/routes/load/window-fetch/correct.svelte
@@ -1,0 +1,18 @@
+<script context="module">
+	/** @type {import('./index').Load} */
+	export async function load({ url, fetch }) {
+		const res = await fetch(`${url.origin}/load/window-fetch/data.json`);
+		const { answer } = await res.json();
+
+		return {
+			props: { answer }
+		};
+	}
+</script>
+
+<script>
+	/** @type {number} */
+	export let answer;
+</script>
+
+<h1>{answer}</h1>
