+++ packages/kit/test/apps/basics/src/routes/load/window-fetch/incorrect.svelte
